Move Dropzone up to top level
